+++ programs/dibio.c
@@ -29,6 +29,7 @@
 #include <errno.h>          /* errno */
 #include <assert.h>
 
+#include "timefn.h"         /* UTIL_time_t, UTIL_clockSpanMicro, UTIL_getTime */
 #include "mem.h"            /* read */
 #include "error_private.h"
 #include "dibio.h"
